+++ src/common/components/filterButton/types.ts
@@ -1,0 +1,13 @@
+export type FilterType =
+  | 'category'
+  | 'hobbyType'
+  | 'date'
+  | 'dateType'
+  | 'division'
+  | 'place'
+  | 'publisher'
+  | 'target'
+  | 'text'
+  | 'minAge'
+  | 'maxAge'
+  | 'exactAge';
