+++ src/domain/eventSearch/filterSummary/TextFilter.tsx
@@ -1,8 +1,7 @@
 import React from 'react';
 
-import FilterButton, {
-  FilterType,
-} from '../../../common/components/filterButton/FilterButton';
+import FilterButton from '../../../common/components/filterButton/FilterButton';
+import { FilterType } from '../../../common/components/filterButton/types';
 
 interface Props {
   onRemove: (value: string, type: FilterType) => void;
